new memory allocation implemented

+++ BaldLionEngine/src/BaldLion/Rendering/Platform/OpenGL/OpenGLTextureCubemap.cpp
@@ -17,7 +17,7 @@ namespace BaldLion
 			glGenTextures(1, &m_rendererID);
 			glBindTexture(GL_TEXTURE_CUBE_MAP, m_rendererID);
 
-			for (int i = 0; i < 6; ++i)
+			for (size_t i = 0; i < 6; ++i)
 			{
 				int width, height, channels;
 				stbi_uc* data = nullptr;
